refactor: V7 content_stats watcher_count 타입 수정

+++ src/main/resources/db/migration/V7__restructure_content_stats.sql
@@ -8,7 +8,7 @@ CREATE TABLE content_stats
     id         UUID PRIMARY KEY,
     review_count  INTEGER          NOT NULL DEFAULT 0,
     rating_sum    DOUBLE PRECISION NOT NULL DEFAULT 0,
-    watcher_count INTEGER          NOT NULL DEFAULT 0,
+    watcher_count BIGINT           NOT NULL DEFAULT 0,
     created_at TIMESTAMPTZ  NOT NULL,
     updated_at TIMESTAMPTZ  NOT NULL
 );
